Restructuring frontend
