+++ frontend/src/App.js
@@ -8,6 +8,9 @@ import NetworkGraph from './NetworkGraph'; // Import NetworkGraph
 import { OpenStreetMapProvider } from 'leaflet-geosearch'; // Corrected Geocoding provider
 import 'leaflet-geosearch/dist/geosearch.css'; // Import Geocoding CSS
 import 'react-vertical-timeline-component/style.min.css'; // Keep existing imports
+import {agentSearch} from './artistSearch';
+import {geocodeLocations} from './geocodeHelper';
+import {artworkSearch} from './artworkSearch';
 
 // Fix for default marker icon issue with Webpack
 import L from 'leaflet';
@@ -48,335 +51,10 @@ function App() {
     };
   }, []);
 
-  // Helper function to extract the earliest year for sorting
-  const parseDateForSort = (dateString) => {
-    if (!dateString) return Infinity; // Place items without dates at the end
-
-    // Handle ranges like "Early 1930s" or "1930s-1940s"
-    // Ignores optional leading text like "Early ", "Late " etc.
-    const rangeMatch = dateString.match(/(?:\b\w+\s+)?(\d{4})s?(?:-(\d{4})s?)?$/);
-    if (rangeMatch) {
-      return parseInt(rangeMatch[1], 10); // Takes the first year in a range
-    }
-
-    // Handle single year or more specific dates, ignoring leading text
-    const yearMatch = dateString.match(/\b(\d{4})\b/);
-    if (yearMatch) {
-      return parseInt(yearMatch[1], 10);
-    }
-
-    // If no year is found, treat as undated
-    return Infinity;
-  };
-
   // --- Geocoding Effect ---
   useEffect(() => {
-    const geocodeLocations = async () => {
-      if (!timelineData || timelineData.length === 0) {
-        console.log("[GEOCODE] timelineData empty, clearing markers.");
-        setMapMarkers([]);
-        return;
-      }
-
-      console.log("[GEOCODE] Processing timelineData:", JSON.parse(JSON.stringify(timelineData))); // Deep copy for logging
-
-      const provider = new OpenStreetMapProvider();
-      // Store promises and their corresponding original items
-      const geocodingRequests = []; // Array of { promise, originalItem }
-      const processedMarkers = []; // Markers with lat/lon already
-
-      timelineData.forEach(item => {
-        if (item.latitude != null && item.longitude != null) {
-          // If lat/lon exist, use them directly
-          // Ensure key includes something unique even if title is missing
-          processedMarkers.push({ ...item, key: `${item.latitude}-${item.longitude}-${item.title || 'no-title'}` }); 
-        } else if (item.location_name) {
-          // If location exists, create a geocoding promise
-          geocodingRequests.push({
-            promise: provider.search({ query: item.location_name }),
-            originalItem: item
-          });
-        } else if (item.latitude == null && item.longitude == null && !item.location_name) {
-          console.warn("[GEOCODE] Item lacks coordinates and location_name:", item);
-        }
-        // Ignore items with neither coordinates nor location
-      });
-
-      // Extract just the promises for Promise.allSettled
-      const promises = geocodingRequests.map(req => req.promise);
-      // Wait for all geocoding requests to settle
-      const settledResults = await Promise.allSettled(promises);
-
-      // Process settled results using the index to map back to the original item
-      settledResults.forEach((result, index) => {
-        const { originalItem } = geocodingRequests[index]; // Get the corresponding original item
-
-        if (result.status === 'fulfilled') {
-          const providerResultsArray = result.value; // Direct result from provider.search
-          console.log(`[GEOCODE] Raw result for '${originalItem.location_name}':`, JSON.parse(JSON.stringify(providerResultsArray))); // Log raw result
-          if (providerResultsArray && providerResultsArray.length > 0) {
-            const geoResult = providerResultsArray[0]; // Use the first result
-            console.log(`[GEOCODE] Geocoded '${originalItem.location_name}' to Lat: ${geoResult.y}, Lon: ${geoResult.x}`);
-            processedMarkers.push({
-              ...originalItem,
-              latitude: geoResult.y,
-              longitude: geoResult.x,
-              // Use index in key for geocoded items to ensure uniqueness
-              key: `${geoResult.y}-${geoResult.x}-${originalItem.title || 'no-title'}-${index}` 
-            });
-          } else {
-            console.warn(`[GEOCODE] No results found for '${originalItem.location_name}'.`);
-          }
-        } else { // result.status === 'rejected'
-          console.error(`[GEOCODE] Geocoding FAILED for '${originalItem.location_name}':`, result.reason);
-        }
-      });
-
-      console.log("[GEOCODE] Final processedMarkers before state update:", JSON.parse(JSON.stringify(processedMarkers)));
-      setMapMarkers(processedMarkers);
-    };
-
-    geocodeLocations();
-
+    geocodeLocations({ timelineData, setMapMarkers });
   }, [timelineData]); // Rerun when timelineData changes
-
-  const agentSearch = async () => {
-    setIsLoading(true);
-    setLoadingTime(0);
-    // Clear previous timer if any
-    if (timerRef.current) {
-      clearInterval(timerRef.current);
-    }
-    // Start new timer
-    timerRef.current = setInterval(() => {
-      setLoadingTime(prevTime => prevTime + 1);
-    }, 1000);
-
-    setTimelineData([]);
-    setNetworkData([]); // Clear network data on new search
-    setError(null);
-    setActiveTimelineScope(''); // Clear active scope indicator
-    setGenreResult(''); // Reset genre result
-
-    try {
-      const response = await fetch(
-        `http://localhost:8080/api/agent?artistName=${artistName}&context=${scope}`
-      );
-
-      if (!response.ok) throw new Error("Error performing AI search");
-
-      const data = await response.json();
-
-      if (data.error) {
-        setError(data.error);
-        // Ensure data arrays are cleared on error
-        setTimelineData([]); 
-        setNetworkData([]);
-        setGenreResult('');
-      } 
-      // --- Conditional Data Handling --- 
-      else if (scope === 'political-events' && data.timelineEvents) {
-        console.log("Received timelineEvents for Political Events:", data.timelineEvents); 
-        console.log('[AGENT_SEARCH] Data before setTimelineData:', JSON.parse(JSON.stringify(data.timelineEvents)));
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Political Events Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data when timeline is loaded
-        setActiveTimelineScope(scope); // Update active scope on success
-      }
-
-      else if (scope === 'art-movements' && data.timelineEvents) { 
-        console.log("Received timelineEvents for Art Movements:", data.timelineEvents);
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-        
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Art Movements Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data
-        setActiveTimelineScope(scope); // Update active scope
-      }
-
-      else if (scope === 'artist-network' && data.networkData) {
-        console.log("Received networkData:", data.networkData); 
-        setNetworkData(data.networkData);
-        setTimelineData([]); // Clear timeline data when network is loaded
-      }
-
-      else if (scope === 'personal-events' && data.timelineEvents) {
-        console.log("Received timelineEvents for Personal Events:", data.timelineEvents); 
-        console.log('[AGENT_SEARCH] Data before setTimelineData:', JSON.parse(JSON.stringify(data.timelineEvents)));
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Personal Events Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data when timeline is loaded
-        setActiveTimelineScope(scope); // Update active scope on success
-      }
-
-      else if (scope === 'economic-events' && data.timelineEvents) {
-        console.log("Received timelineEvents for Economic Events:", data.timelineEvents); 
-        console.log('[AGENT_SEARCH] Data before setTimelineData:', JSON.parse(JSON.stringify(data.timelineEvents)));
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Economic Events Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data when timeline is loaded
-        setActiveTimelineScope(scope); // Update active scope on success
-      }
-
-      else if (scope === 'genre' && data.timelineEvents) {
-        console.log("Received timelineEvents for Genre:", data.timelineEvents); 
-        console.log('[AGENT_SEARCH] Data before setTimelineData:', JSON.parse(JSON.stringify(data.timelineEvents)));
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Genre Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data when timeline is loaded
-        setActiveTimelineScope(scope); // Update active scope on success
-      }
-
-      else if (scope === 'medium' && data.timelineEvents) {
-        console.log("Received timelineEvents for Medium:", data.timelineEvents); 
-        console.log('[AGENT_SEARCH] Data before setTimelineData:', JSON.parse(JSON.stringify(data.timelineEvents)));
-        let transformedData = (data.timelineEvents || []).map(event => ({
-          title: event.date, 
-          cardTitle: event.event_title,
-          location_name: event.location_name, // Include location_name
-          cardDetailedText: event.detailed_summary + (event.source_url ? `\n\n[Source](${event.source_url})` : ""), 
-          
-          // images to show up in timeline where needed
-          media: event.artwork_image_url ? {
-            type: "IMAGE",
-            source: {
-              url: event.artwork_image_url
-            }
-          } : undefined,
-          latitude: event.latitude,
-          longitude: event.longitude
-        }));
-
-        // Sort chronologically based on the earliest year found in the date
-        transformedData.sort((a, b) => parseDateForSort(a.title) - parseDateForSort(b.title));
-        console.log('[AGENT_SEARCH] Sorted Medium Data:', JSON.parse(JSON.stringify(transformedData)));
-
-        setTimelineData(transformedData);
-        setNetworkData([]); // Clear network data when timeline is loaded
-        setActiveTimelineScope(scope); // Update active scope on success
-      }
-
-      else {
-        // Handle cases where data is missing for the expected scope
-        console.warn(`Data key ('${scope === 'political-events' ? 'timelineEvents' : 'networkData'}') not found in response for scope '${scope}'.`);
-        setError(`Received response, but no valid data found for the selected scope.`);
-        setTimelineData([]);
-        setNetworkData([]);
-        setGenreResult('');
-      }
-      // ----------------------------------
-    } catch (err) {
-      setError(err.message || "Failed to perform AI search");
-      setTimelineData([]);
-      setNetworkData([]); // Clear network data on fetch error too
-      setGenreResult('');
-    }
-    finally {
-      setIsLoading(false);
-      if (timerRef.current) {
-        clearInterval(timerRef.current);
-        timerRef.current = null;
-      }
-    }
-  }
 
   const getListTitle = (currentScope) => { // Modified to accept scope argument
     switch(currentScope) {
@@ -448,7 +126,39 @@ function App() {
           </select>
         </div>
 
-        <button onClick={agentSearch}>Search</button>
+        <button onClick={() => {
+          if(!artworkTitleOn || artworkTitle === "" || artworkTitle.toLowerCase() === "untitled") {
+            agentSearch({
+              artistName,
+              scope,
+              setIsLoading,
+              setLoadingTime,
+              timerRef,
+              setTimelineData,
+              setNetworkData,
+              setError,
+              setActiveTimelineScope,
+              setGenreResult
+            });
+          } else {
+            // TODO implement search by artwork
+            artworkSearch({
+              artistName,
+              artworkTitle,
+              scope,
+              setIsLoading,
+              setLoadingTime,
+              timerRef,
+              setTimelineData,
+              setNetworkData,
+              setError,
+              setActiveTimelineScope,
+              setGenreResult,
+              setArtworkTitle
+            });
+        }
+        }}>
+          Search</button>
       </div>
 
       {/* Loading Indicator */}
