+++ frontend/src/artworkSearch.js
@@ -1,0 +1,18 @@
+export const artworkSearch = async({
+                                       artistName,
+                                       artworkTitle,
+                                       scope,
+                                       setIsLoading,
+                                       setLoadingTime,
+                                       timerRef,
+                                       setTimelineData,
+                                       setNetworkData,
+                                       setError,
+                                       setActiveTimelineScope,
+                                       setGenreResult,
+                                       setArtworkTitle
+                                   }) => {
+    //TODO implement this
+    setArtworkTitle("Performing search by artwork!");
+
+}
